Hide "reload" when building for wasm
On current wasm targets (both browser and VSCode) the reload
doesn't work. In the browser it's clear (we need a new "upload")
on VSCode we might be able to deal with it, but doesn't work now.

Hide "reload" in both menu bar as well as menu bar and command
prompt to prevent confusion/wrong assumptions by users.